+++ app/Helpers/zones.php
@@ -13,8 +13,7 @@ function getZones(): LengthAwarePaginator
 
     $directions = (array) request()->input('directions', []);
 
-    /** @phpstan-ignore-next-line */
-    $perPage = (int) request()->integer('perPage', 10);
+    $perPage = request()->integer('perPage', 10);
 
     $zones = Zone::search($search);
